more directory and code refactoring
